moved test config files
